renamed rbuilders package and its classes
