+++ apex-squid/src/main/java/org/fundacionjala/enforce/sonarqube/apex/api/grammar/builders/KeywordRules.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.fundacionjala.enforce.sonarqube.apex.api.grammar.rbuilders;
+package org.fundacionjala.enforce.sonarqube.apex.api.grammar.builders;
 
 import org.sonar.sslr.grammar.LexerfulGrammarBuilder;
 import static org.fundacionjala.enforce.sonarqube.apex.api.ApexKeyword.SHARING;
@@ -35,7 +35,7 @@ import static org.fundacionjala.enforce.sonarqube.apex.api.grammar.ApexGrammarRu
  *
  * @author kevin_titichoca
  */
-public class KeywordRulesBuilder {
+public class KeywordRules {
 
     public static void create(LexerfulGrammarBuilder grammarBuilder) {
         keyword(grammarBuilder);
