+++ apex-squid/src/main/java/org/fundacionjala/enforce/sonarqube/apex/api/grammar/builders/TypeRules.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.fundacionjala.enforce.sonarqube.apex.api.grammar.rbuilders;
+package org.fundacionjala.enforce.sonarqube.apex.api.grammar.builders;
 
 import org.sonar.sslr.grammar.LexerfulGrammarBuilder;
 import static org.fundacionjala.enforce.sonarqube.apex.api.ApexKeyword.ABSTRACT;
@@ -66,7 +66,7 @@ import static org.fundacionjala.enforce.sonarqube.apex.api.grammar.ApexGrammarRu
  *
  * @author kevin_titichoca
  */
-public class TypeRuleBuilder {
+public class TypeRules {
 
     public static void create(LexerfulGrammarBuilder grammarBuilder) {
         typeSpecifier(grammarBuilder);
